Using decimals consistently system wide

+++ StrategyTemplate/ForYou/GraphBar.cs
@@ -8,14 +8,14 @@ namespace StrategyTemplate.ForYou
 {
     public struct GraphBar
     {
-        public double OpeningPrice {get; private set;}
-        public double ClosingPrice {get; private set;}
-        public double PriceHigh {get; private set;}
-        public double PriceLow {get; private set;}
-        public double Volume {get; private set;}
+        public decimal OpeningPrice {get; private set;}
+        public decimal ClosingPrice {get; private set;}
+        public decimal PriceHigh {get; private set;}
+        public decimal PriceLow {get; private set;}
+        public decimal Volume {get; private set;}
         public DateTime TimeStampUTC {get; private set;}
 
-        public GraphBar(double open, double close, double high, double low, double volume, DateTime timeStamp)
+        public GraphBar(decimal open, decimal close, decimal high, decimal low, decimal volume, DateTime timeStamp)
         {
             OpeningPrice = open;
             ClosingPrice = close;
